% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_time_weighted_token_balance.R
\name{address_time_weighted_token_balance}
\alias{address_time_weighted_token_balance}
\title{Time Weighted Address Token Balance}
\usage{
address_time_weighted_token_balance(
  token_address,
  min_tokens = 1e-04,
  block_min = 0,
  block_max,
  amount_weighting = TRUE,
  api_key
)
}
\arguments{
\item{token_address}{ERC20 token contract address to assess balance.}

\item{min_tokens}{Minimum amount of tokens acknowledged. Already decimal adjusted, useful to
ignoring dust balances.}

\item{block_min}{Initial block to start scoring balances over time, default 0 (genesis block).}

\item{block_max}{The block height to assess balance at (for reproducibility).}

\item{amount_weighting}{Weight by amounts held across time, default TRUE. If FALSE, it treats
all amounts as binary. Person had at least `min_tokens` at a block or they did not. To NOT
weight by time, use `address_token_balance()` instead of this function.}

\item{api_key}{Flipside Crypto ShroomDK API Key to create and access SQL queries.}
}
\value{
A data frame
}
\description{
Time Weighted Address Token Balance
}
\examples{
\dontrun{
address_time_weighted_token_balance(
token_address = tolower("0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984"), #UNI token
 min_tokens = 0.01,
 block_min = 10000000,
 block_max = 15000000,
 amount_weighting = TRUE,
 api_key = readLines("api_key.txt")
)
}
}
