% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_token_accumulate.R
\name{address_token_accumulate}
\alias{address_token_accumulate}
\title{Address Token Accumulation}
\usage{
address_token_accumulate(
  token_address,
  block_min = 0,
  block_max,
  min_tokens = 0.0001,
  api_key
)
}
\arguments{
\item{token_address}{ERC20 token contract address to assess change in balance.}

\item{block_min}{Initial block to start scoring balances over time, default 0 (genesis block).}

\item{block_max}{The block height to assess balance at (for reproducibility).}

\item{min_tokens}{Minimum amount of tokens acknowledged. Already decimal adjusted, useful to
ignoring dust balances. Default 0.0001. Use -Inf to include net sellers (but note: API max is 1M rows.).}

\item{api_key}{Flipside Crypto ShroomDK API Key to create and access SQL queries.}
}
\value{
A data frame of the form:\tabular{lc}{
    \tab  \cr
   ADDRESS \tab The EOA or contract that holds the balance \cr
   TOKEN_ADDRESS \tab ERC20 address provided \cr
   NET_CHANGE \tab amount of tokens accumulated between block_min and block_max \cr
}
}
\description{
Net accumulation of a token between two block heights. By default excludes net sellers.
Alice net gained 200 UNI between blocks 10,000,000 and 15,000,000.
}
\examples{
\dontrun{
address_token_accumulate(
token_address = tolower("0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984"), #UNI token
 block_min = 10000000,
 block_max = 15000000,
 min_tokens = 1,
 api_key = readLines("api_key.txt")
)
}
}
