% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_token_transfer_volume.R
\name{address_transfer_volume}
\alias{address_transfer_volume}
\title{Address Token Transfer Volume}
\usage{
address_transfer_volume(
  token_address,
  block_min = 0,
  block_max,
  min_tokens = 1,
  api_key
)
}
\arguments{
\item{token_address}{ERC20 token contract address to assess change in balance.}

\item{block_min}{Initial block to start scoring balances over time, default 0 (genesis block).}

\item{block_max}{The block height to assess balance at (for reproducibility).}

\item{min_tokens}{Minimum amount of tokens acknowledged. Already decimal adjusted, useful for
ignoring dust balances. Default 1.}

\item{api_key}{Flipside Crypto ShroomDK API Key to create and access SQL queries.}
}
\value{
A data frame of the form:\tabular{lc}{
    \tab  \cr
   ADDRESS \tab The EOA or contract with the volume \cr
   TOKEN_ADDRESS \tab ERC20 address provided \cr
   VOLUME \tab amount of tokens transferred/traded between block_min and block_max \cr
   ADDRESS_TYPE \tab If ADDRESS is known to be 'contract address' or 'gnosis safe address'. \cr
   If neither it is assumed to be an 'EOA'. Some EOAs may have a balance but have never initiated a transaction \tab  \cr
   These are noted as 'EOA-0tx' Note: contracts, including gnosis safes may not have consistent owners across different EVM chains. \tab  \cr
}
}
\description{
Total transfer volume of a token between two block heights. Ignores direction of transfer.
Alice traded/transferred a total of 2,000 UNI between blocks 10,000,000 and 15,000,000.
}
\examples{
\dontrun{
address_transfer_volume(
token_address = tolower("0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984"), #UNI token
 block_min = 10000000,
 block_max = 15000000,
 min_tokens = 10,
 api_key = readLines("api_key.txt")
)
}
}
