% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_int_to_hex.R
\name{bigint_to_hex}
\alias{bigint_to_hex}
\title{Big Int to Hex}
\usage{
bigint_to_hex(bigint)
}
\arguments{
\item{bigint}{vector of large integers as characters.}
}
\value{
vector of hex with leading 0x and leading 0s to ensure at least 66 length.
This 66 length is to best fit Flipside Crypto ENS data.
}
\description{
Converts extremely large numbers into HEX to make them workable. Most specifically
used for NFTs with millions of millions of token IDs. see `get_ens()` as a main example.
}
\examples{
\dontrun{
token_id = "41104824783848331047501863836715107956672917465157448818057950770477717896101"
bigint_to_hex(token_i)
}
}
