% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_address_activity.R
\name{get_address_activity}
\alias{get_address_activity}
\title{Get Address Activity}
\usage{
get_address_activity(addresses, block_min = 0, block_max, api_key)
}
\arguments{
\item{addresses}{a character vector of addresses}

\item{block_min}{Initial block to start scoring balances over time, default 0 (genesis block).}

\item{block_max}{The block height to assess balance at (for reproducibility).}

\item{api_key}{Flipside Crypto ShroomDK API Key to create and access SQL queries.}
}
\value{
A data frame of the form:
| |  |
| ------------- |:-------------:|
| ADDRESS       | The EOA or contract that holds the balance |
| NUM_TX        | Number of transactions initiated by ADDRESS between block heights. Note, EOA-0tx and contracts may return null. |
| NUM_DAYS | Number of unique days with a transaction between block heights |
| LAST_TX_DATE | YYYY-MM-DD date (UTC) of last transaction of ADDRESS between block heights  |
}
\description{
Gets basic activity information from a set of addresses: number of transactions,
number of days active, and last transaction date.
}
\examples{
\dontrun{
alist <- c("0x39e856863e5f6f0654a0b87b12bc921da23d06bb",
 "0xd8da6bf26964af9d7eed9e03e53415d37aa96045")
get_address_activity(
 addresses = alist,
 block_min = 10000000,
 block_max = 15000000,
 api_key = readLines("api_key.txt")
 )
}
}
