% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ens.R
\name{get_ens}
\alias{get_ens}
\title{Get Ethereum Name Service
Identifies ENS NFT names of addresses. Note, currently does not subset to current
holder. All historical owners will have the NFT. It is possible non-EOAs have an ENS,
e.g., gnosis safe holders and contracts.}
\usage{
get_ens(addresses, api_key)
}
\arguments{
\item{addresses}{a character vector of addresses}

\item{api_key}{Flipside Crypto ShroomDK API Key to create and access SQL queries}
}
\value{
Data frame of the form:
| |  |
| ------------- |:-------------:|
| ADDRESS  | EOA Address that held the ENS at least once. |
| ENS_NAME | Ethereum Name Service name, traders of ENS NFTs will each show as a historic recipient |
| TOKENID  |  ENS NFT Token ID Number (be aware of BIGINT issue) |
| HEX_TOKENID | ENS NFT Token ID in HEX form (to mitigate BIGINT issue) |
}
\description{
Get Ethereum Name Service
Identifies ENS NFT names of addresses. Note, currently does not subset to current
holder. All historical owners will have the NFT. It is possible non-EOAs have an ENS,
e.g., gnosis safe holders and contracts.
}
\examples{
\dontrun{
get_ens(addresses = "0x39E856863e5F6f0654a0b87B12bc921DA23D06BB",
       api_key = readLines("api_key.txt"))
}
}
