% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_bigint.R
\name{hex_to_bigint}
\alias{hex_to_bigint}
\title{Hex to Big Int}
\usage{
hex_to_bigint(hex)
}
\arguments{
\item{hex}{Hex to convert to big integer. Leading 0s are appropriately ignored.}
}
\value{
vector of big integer in character format.
}
\description{
Converts HEX to extremely large numbers to make them workable. Most specifically
used for NFTs with millions of millions of token IDs. see `get_ens()` as a main example.
}
\examples{
\dontrun{
hex = "0x2aea5a3e13081115c6fcc5503f7ce0d950a8dfc5eaf871"
bigint_to_hex(token_i)
}
}
